[SPARK-38043][SQL] Refactor FileBasedDataSourceSuite and add DataSourceSuite for each data source
